[GREEN-29] Tax refactoring
